extracting ThemeContext to its own dedicated component and file
